First projectile component implementation. One remaining bug: unregularly, artifacts appear when displaying multiple projectiles.

+++ include/components/ColliderComponent.hpp
@@ -45,8 +45,8 @@ public:
         if(tag != "terrain") {
             collider.x = static_cast<int>(transform->position.x);
             collider.y = static_cast<int>(transform->position.y);
-            collider.w = static_cast<int>(transform->width * transform->scale);
-            collider.h = static_cast<int>(transform->height * transform->scale);
+            collider.w = transform->width * transform->scale;
+            collider.h = transform->height * transform->scale;
         }
 
         dstRect.x = collider.x - Game::camera.x;
