+++ include/components/TransformComponent.hpp
@@ -39,8 +39,8 @@ public:
     }
 
     void update() override {
-        position.x += velocity.x * speed;
-        position.y += velocity.y * speed;
+        position.x += static_cast<int>(velocity.x * speed);
+        position.y += static_cast<int>(velocity.y * speed);
     }
 
 private:
